Scenarios: protected areas threshold
