+++ app/layout/scenarios/sidebar/wdpa/categories/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
